requirements.txt
altair
numpy
pandas
pydeck
streamlit==1.42.0
matplotlib
seaborn
yfinance
scipy
plotly
scikit-learn
xgboost
openpyxl
statsmodels
xlrd
xlsxwriter
vectorbt
arch
python-bcb
beautifulsoup4
